feat(design-system): épico 4 — refactor pages to use self-fetching feature sections
Pages (Home, About, Projects) are now thin orchestrators that only import
from ~features/*. Data fetching and rendering logic moved into dedicated
Server Component sections: home/HeroSection, home/ProjectsSection,
contact/ContactSection, about/HeroSection, about/ValuesSection,
about/ExperiencesSection, projects/HeroSection, projects/ProjectsSection.
