+++ apps/web/src/features/about/ExperiencesSection/index.ts
@@ -1,1 +1,2 @@
 export * from './ExperienceCard';
+export { ExperiencesSection } from './ExperiencesSection';
