+++ apps/web/src/features/about/ValuesSection/index.ts
@@ -1,1 +1,2 @@
 export * from './ProfessionalValue';
+export { ValuesSection } from './ValuesSection';
