+++ apps/web/src/features/home/ProjectsSection/index.ts
@@ -1,2 +1,3 @@
 export * from './ProjectCard';
 export * from './ProjectList';
+export { ProjectsSection } from './HomeProjectsSection';
